+++ packages/methods/src/hasDeepErrors/index.ts
@@ -1,0 +1,1 @@
+export * from './hasDeepErrors.ts';